chore: bundle serverless PDF.js build with rolldown
