+++ pdfjs.rolldown.config.ts
@@ -1,0 +1,43 @@
+import { defineConfig } from 'rolldown'
+import { patchPDFJSSource, pdfjsTypes } from './src/pdfjs-serverless/rolldown/plugins'
+
+export default defineConfig({
+  input: 'src/pdfjs-serverless/index.mjs',
+  // The package sets `"sideEffects": false` for the tree-shakeable library
+  // build, but the serverless entry pulls in polyfill/mock modules whose only
+  // job is to mutate global prototypes (`Uint8Array.prototype.toHex`,
+  // `Map.prototype.getOrInsertComputed`, `DOMMatrix`, …). Force them to be
+  // treated as side-effectful so their installs are never tree-shaken away.
+  treeshake: {
+    moduleSideEffects: [
+      { test: /pdfjs-serverless[\\/](mocks|polyfills)\.mjs$/, sideEffects: true },
+    ],
+  },
+  output: {
+    file: 'dist/pdfjs.mjs',
+    format: 'esm',
+    exports: 'auto',
+    // The worker is pulled in via a static import and inlined through the
+    // `__pdfjsWorker__` anchor, so everything must land in a single chunk.
+    codeSplitting: false,
+    sourcemap: false,
+    // PDF.js relies on `Function.prototype.name`/class names at runtime, so the
+    // minifier must preserve them – the oxc equivalent of esbuild's `keepNames`.
+    minify: {
+      compress: {
+        target: 'es2022',
+        keepNames: { function: true, class: true },
+      },
+      mangle: {
+        keepNames: { function: true, class: true },
+      },
+      codegen: {
+        legalComments: 'none',
+      },
+    },
+  },
+  plugins: [
+    patchPDFJSSource(),
+    pdfjsTypes(),
+  ],
+})
